#48: extracting `Config` to separate file
